User profile item component

+++ client/src/components/Items/UsersItemsList.jsx
@@ -2,11 +2,10 @@ import React, { useState, useEffect } from "react";
 import Item from './Item'
 import express from "../../apis/express";
 
-const ItemList = ({session}) => {
+const UsersItemList = ({session}) => {
 
   const [itemList, setItemList] = useState([])
-  const [name, setName] = useState("")
-  const [description, setDescription] = useState("")
+
 
   useEffect(() => {
     getItems()
@@ -17,39 +16,14 @@ const ItemList = ({session}) => {
     setItemList(data)
   }
 
-  const addItem = async (event) => {
-    event.preventDefault()
-    if(session) {
-      const { id: userId } = session.user
-      const { data } = await express.post('/items', {
-        item: {
-          name: name,
-          description: description,
-          lender: userId
-        }
-      })
-    }
-    
-    getItems()
-  }
-
   const deleteItem = async (id) => {
-
     const { data } = await express.delete(`/items/${id}`)
     getItems()
   }
 
-  const borrowItem = async (id) => {
-    if(session) {
-      const { id: userId } = session.user
-      const { data } = await express.put(`/items/${id}/borrow/${userId}`)
-    }
-    getItems()
-  }
 
   const updateItem = async (event, name, description, lender, borrower, id) => {
     event.preventDefault()
-    // const { data } = 
     await express.put(`/items/${id}`, {
       item: {
         name: name,
@@ -61,7 +35,6 @@ const ItemList = ({session}) => {
   }
 
   const uploadImage = async (event, imageTitle, imageFile, id) => {
-    
     event.preventDefault()
     let formData = new FormData()
     formData.append('title', imageTitle)
@@ -77,25 +50,26 @@ const ItemList = ({session}) => {
   }
 
   const renderedList = itemList.map(item => {
-    return (
-      <Item 
-        key={item._id} 
-        item={item}
-        deleteItem={deleteItem}
-        updateItem={updateItem}
-        uploadImage={uploadImage}
-        borrowItem={borrowItem}
-      />
-    )
+        const user = JSON.parse(localStorage.getItem('user'));
+        const userId = user.user.id;
+    
+    if (item.lender._id == userId) {
+      return (
+        <Item 
+          key={item._id} 
+          item={item}
+          deleteItem={deleteItem}
+          updateItem={updateItem}
+          uploadImage={uploadImage}
+        />
+      )
+
+    }
   })
 
   return (
     <div>
-      <form onSubmit={addItem}>
-        <input onChange={(event) => setName(event.target.value)} value={name} type="text" name="name" required/>
-        <input onChange={(event) => setDescription(event.target.value)} value={description} type="text" name="description" required/>
-        <button>Add Item</button>
-      </form>
+
       <div className="ui two stackable cards">
         {renderedList}
       </div>
@@ -105,4 +79,4 @@ const ItemList = ({session}) => {
 
 }
 
-export default ItemList
+export default UsersItemList
